Bug 1667877: Remove unnecessary "secure revision without bug" code
The only case in which a secure revision would not have a bug is when a
revision is first created and hasn't been processed by "phab-bot" yet.
Since emails shouldn't be submitted for a revision until it's been
checked by "phab-bot", we don't need to worry about this case.

Therefore, the code can be marginally simplified.