Bug 1672239: Improve resiliency towards render/send failures
Drastically reduces the number of cases where we'd "lose" an email
by falling back to a "minimal" variant, if possible.

Note that a tradeoff was made here such that it's still possible
to lose an email if it can't be rendered or sent, even in
minimal form:

IMHO, an ideal solution would have been to block the queue if a
minimal email cannot be rendered or sent. However, due to
trickiness in tracking state to avoid double-sends (what if the
full context fails to be sent to half the recipients, but we fail
to render the minimal context for those unfortunate few?), it
seemed more effective in the short-term to still skip emails
in these scenarios, but to loudly proclaim the issue in both
stats and Sentry.

As discussed in the bug, there's future work around managing state
so that emails don't become inaccurate if they're sent late. In
the same way, there's possible future work to completely avoid
dropped emails.