Bug 1633922: Use ellipsis on long file paths for comments
Now that we're showing the absolute path of files that are
commenting on, we should gracefully handle the case where
there's not enough room to show the entire path on a single
line.

I've opted to put the ellipsis at the beginning of the path
instead of the end because the affected file is the most
important part.