Bug 1669586: Split "closed" from "landed" event emails
There's two different workflows for closing revisions, with different
contexts:
* A user chooses the "close" action within Phabricator, and perhaps
  leaves some comments.
* Phabricator detects that the code associated with the revision has
  landed in the actual VCS repo. In this case, it's useful for email
  recipients to see the revision hash and some relevant links.

This patch splits up the existing "Landed" event to cover both of these
workflows.

Note for reviewers: the "closed" templates are copies of the old
"landed" templates, and the only piece that has been modified is the
emoji (to "chequered_flag"). You shouldn't need to manually review the
whole templates.