Bug 1711964: Consider most AWS errors as not caused by the outgoing mail
We only want to bubble up "permanent error" and skip an outgoing message
if the message itself is causing the failure.
However, all SES errors that I'm currently aware of either directly
or potentially caused by configuration issues and *not* the message.

My original assumption of `MessageRejected` referring to issues with
the message was proved incorrect by this official document:
https://aws.amazon.com/premiumsupport/knowledge-center/ses-554-400-message-rejected-error/