Bug 1714175: Handle UTF-8 in "from" email header
Now that we have "user_name (real_name) <mozphab@mozilla.com>" in
the "From" header, we need to handle the case where `user_name` or
`real_name` have non-ASCII unicode values.

Python has the tools to handle this, but surprisingly it's necessary
to explicitly only encode the potentially-utf8 section - if the
email address in the `<...>` is encoded, then SES raises an error
because it wants it in plaintext. So, that's why only a substring
of `From` has been encoded as a `Header()`.