Bug 1674796: Use threading headers
Provide four threading headers: `In-Reply-To`, `References`,
`Thread-Topic` and `Thread-Index`.

The implementation of headers here is verify similar to the behaviour of
Phabricator revision email headers.

Note that an "imaginary" top-level message ID was used instead of
referring to the "first message" in each thread because each user may
have a different "first message", and tracking that information would've
been nontrivial. Fortunately, email clients are resilient to grouping by
"imaginary parent", so this tradeoff should be acceptable.